move pytest fixtures to conftest.py
