+++ tests/components/vizio/test_config_flow.py
@@ -1,6 +1,4 @@
 """Tests for Vizio config flow."""
-
-from asynctest import patch
 import pytest
 import voluptuous as vol
 
@@ -39,55 +37,54 @@ from .const import (
 
 from tests.common import MockConfigEntry
 
-
-@pytest.fixture(name="vizio_connect")
-def vizio_connect_fixture():
-    """Mock valid vizio device and entry setup."""
-    with patch(
-        "homeassistant.components.vizio.config_flow.VizioAsync.validate_ha_config",
-        return_value=True,
-    ), patch(
-        "homeassistant.components.vizio.config_flow.VizioAsync.get_unique_id",
-        return_value=UNIQUE_ID,
-    ):
-        yield
-
-
-@pytest.fixture(name="vizio_bypass_setup")
-def vizio_bypass_setup_fixture():
-    """Mock component setup."""
-    with patch("homeassistant.components.vizio.async_setup_entry", return_value=True):
-        yield
+# @pytest.fixture(name="vizio_connect")
+# def vizio_connect_fixture():
+#     """Mock valid vizio device and entry setup."""
+#     with patch(
+#         "homeassistant.components.vizio.config_flow.VizioAsync.validate_ha_config",
+#         return_value=True,
+#     ), patch(
+#         "homeassistant.components.vizio.config_flow.VizioAsync.get_unique_id",
+#         return_value=UNIQUE_ID,
+#     ):
+#         yield
 
 
-@pytest.fixture(name="vizio_bypass_update")
-def vizio_bypass_update_fixture():
-    """Mock component update."""
-    with patch(
-        "homeassistant.components.vizio.media_player.VizioAsync.can_connect",
-        return_value=True,
-    ), patch("homeassistant.components.vizio.media_player.VizioDevice.async_update"):
-        yield
+# @pytest.fixture(name="vizio_bypass_setup")
+# def vizio_bypass_setup_fixture():
+#     """Mock component setup."""
+#     with patch("homeassistant.components.vizio.async_setup_entry", return_value=True):
+#         yield
 
 
-@pytest.fixture(name="vizio_guess_device_type")
-def vizio_guess_device_type_fixture():
-    """Mock vizio async_guess_device_type function."""
-    with patch(
-        "homeassistant.components.vizio.config_flow.async_guess_device_type",
-        return_value="speaker",
-    ):
-        yield
+# @pytest.fixture(name="vizio_bypass_update")
+# def vizio_bypass_update_fixture():
+#     """Mock component update."""
+#     with patch(
+#         "homeassistant.components.vizio.media_player.VizioAsync.can_connect",
+#         return_value=True,
+#     ), patch("homeassistant.components.vizio.media_player.VizioDevice.async_update"):
+#         yield
 
 
-@pytest.fixture(name="vizio_cant_connect")
-def vizio_cant_connect_fixture():
-    """Mock vizio device cant connect."""
-    with patch(
-        "homeassistant.components.vizio.config_flow.VizioAsync.validate_ha_config",
-        return_value=False,
-    ):
-        yield
+# @pytest.fixture(name="vizio_guess_device_type")
+# def vizio_guess_device_type_fixture():
+#     """Mock vizio async_guess_device_type function."""
+#     with patch(
+#         "homeassistant.components.vizio.config_flow.async_guess_device_type",
+#         return_value="speaker",
+#     ):
+#         yield
+
+
+# @pytest.fixture(name="vizio_cant_connect")
+# def vizio_cant_connect_fixture():
+#     """Mock vizio device cant connect."""
+#     with patch(
+#         "homeassistant.components.vizio.config_flow.VizioAsync.validate_ha_config",
+#         return_value=False,
+#     ):
+#         yield
 
 
 async def test_unload(
